Comment performance
fix rebase

Improvements, remove global state

comment separate file
