+++ packages/atlas/src/api/queries/comments.graphql
@@ -15,12 +15,12 @@ query GetComments(
   }
 }
 
-query GetComment($where: CommentWhereUniqueInput!, $memberId: ID, $videoId: ID) {
-  commentReactions(where: { member: { id_eq: $memberId }, video: { id_eq: $videoId } }, limit: 1000) {
+query GetComment($commentID: ID!) {
+  commentReactions(where: { comment: { id_eq: $commentID } }, limit: 1000) {
     reactionId
     commentId
   }
-  commentByUniqueInput(where: $where) {
+  commentByUniqueInput(where: { id: $commentID }) {
     ...CommentFields
   }
 }
