+++ packages/atlas/src/hooks/useReactionTransactions.ts
@@ -1,5 +1,5 @@
 import { useApolloClient } from '@apollo/client'
-import { useCallback, useState } from 'react'
+import { useCallback } from 'react'
 
 import {
   GetCommentsDocument,
@@ -17,23 +17,7 @@ import { ConsoleLogger } from '@/utils/logs'
 export const useReactionTransactions = () => {
   const { activeMemberId } = useUser()
   const { joystream, proxyCallback } = useJoystream()
-  // stores the isProcessing state of comment inputs and is indexed by commentId's
-  const [commentInputIsProcessingCollection, setCommentInputIsProcessingCollection] = useState(new Set<string>())
-  const setCommentInputIsProcessing = ({ commentInputId, value }: { commentInputId: string; value: boolean }) => {
-    setCommentInputIsProcessingCollection((processing) => {
-      if (value) {
-        processing.add(commentInputId)
-      } else {
-        processing.delete(commentInputId)
-      }
-      return new Set(processing)
-    })
-  }
-
-  const [videoReactionProcessing, setVideoReactionProcessing] = useState(false)
-
   const handleTransaction = useTransaction()
-  const [processingCommentReactionId, setProcessingCommentReactionId] = useState<string | null>(null)
 
   const client = useApolloClient()
 
@@ -45,46 +29,12 @@ export const useReactionTransactions = () => {
     [client]
   )
 
-  const reactToComment = useCallback(
-    async (commentId: string, reactionId: ReactionId) => {
-      if (!joystream || !activeMemberId) {
-        ConsoleLogger.error('No joystream instance')
-        return
-      }
-
-      setProcessingCommentReactionId(commentId + `-` + reactionId.toString())
-
-      return handleTransaction({
-        txFactory: async (updateStatus) =>
-          (await joystream.extrinsics).reactToVideoComment(
-            activeMemberId,
-            commentId,
-            reactionId,
-            proxyCallback(updateStatus)
-          ),
-        minimized: {
-          signErrorMessage: 'Failed to react to comment',
-        },
-        onTxSync: async () => {
-          await refetchComments()
-          setProcessingCommentReactionId(null)
-        },
-        onError: () => {
-          setProcessingCommentReactionId(null)
-        },
-      })
-    },
-    [activeMemberId, handleTransaction, joystream, proxyCallback, refetchComments]
-  )
-
   const addComment = useCallback(
     async ({
-      commentInputId,
       parentCommentId,
       videoId,
       commentBody,
     }: {
-      commentInputId: string
       parentCommentId?: string
       videoId: string
       commentBody: string
@@ -93,7 +43,6 @@ export const useReactionTransactions = () => {
         ConsoleLogger.error('no joystream or active member')
         return
       }
-      setCommentInputIsProcessing({ commentInputId, value: true })
 
       let newCommentId: string | undefined
 
@@ -110,8 +59,6 @@ export const useReactionTransactions = () => {
           ),
         onTxSync: async ({ transactionHash }) => {
           const refetchResult = await refetchComments()
-          setCommentInputIsProcessing({ commentInputId, value: false })
-
           const { data } = refetchResult[0]
 
           newCommentId = data?.videoCommentsConnection.edges.find(
@@ -119,9 +66,6 @@ export const useReactionTransactions = () => {
             // update this once QN supports getting ID directly from the status query
             (edge) => edge.node.commentcreatedeventcomment?.[0].inExtrinsic === transactionHash
           )?.node.id
-        },
-        onError: () => {
-          setCommentInputIsProcessing({ commentInputId, value: false })
         },
         minimized: {
           signErrorMessage: 'Failed to post video comment',
@@ -132,25 +76,46 @@ export const useReactionTransactions = () => {
     [activeMemberId, handleTransaction, joystream, proxyCallback, refetchComments]
   )
 
+  const reactToComment = useCallback(
+    async (commentId: string, reactionId: ReactionId) => {
+      if (!joystream || !activeMemberId) {
+        ConsoleLogger.error('No joystream instance')
+        return
+      }
+
+      return handleTransaction({
+        txFactory: async (updateStatus) =>
+          (await joystream.extrinsics).reactToVideoComment(
+            activeMemberId,
+            commentId,
+            reactionId,
+            proxyCallback(updateStatus)
+          ),
+        minimized: {
+          signErrorMessage: 'Failed to react to comment',
+        },
+        onTxSync: async () => {
+          await refetchComments()
+        },
+      })
+    },
+    [activeMemberId, handleTransaction, joystream, proxyCallback, refetchComments]
+  )
+
   const updateComment = useCallback(
     async ({ commentId, videoId, commentBody }: { commentId: string; videoId: string; commentBody: string }) => {
       if (!joystream || !activeMemberId || !videoId) {
         ConsoleLogger.error('no joystream or active member')
-        return
+        return false
       }
-      setCommentInputIsProcessing({ commentInputId: commentId, value: true })
 
-      await handleTransaction({
+      return await handleTransaction({
         txFactory: async (updateStatus) =>
           (
             await joystream.extrinsics
           ).editVideoComment(activeMemberId, commentId, commentBody, proxyCallback(updateStatus)),
         onTxSync: async () => {
           await refetchComments()
-          setCommentInputIsProcessing({ commentInputId: commentId, value: false })
-        },
-        onError: () => {
-          setCommentInputIsProcessing({ commentInputId: commentId, value: false })
         },
         minimized: {
           signErrorMessage: 'Failed to udpate video comment',
@@ -226,12 +191,10 @@ export const useReactionTransactions = () => {
     (videoId: string, reaction: VideoReaction) => {
       if (!joystream || !activeMemberId) {
         ConsoleLogger.error('No joystream instance')
-        return
+        return Promise.reject(false)
       }
 
-      setVideoReactionProcessing(true)
-
-      handleTransaction({
+      return handleTransaction({
         txFactory: async (updateStatus) =>
           (await joystream.extrinsics).reactToVideo(activeMemberId, videoId, reaction, proxyCallback(updateStatus)),
         minimized: {
@@ -239,10 +202,6 @@ export const useReactionTransactions = () => {
         },
         onTxSync: async () => {
           await refetchVideo()
-          setVideoReactionProcessing(false)
-        },
-        onError: async () => {
-          setVideoReactionProcessing(false)
         },
       })
     },
@@ -250,14 +209,11 @@ export const useReactionTransactions = () => {
   )
 
   return {
-    reactToComment,
     addComment,
+    reactToComment,
     deleteComment,
     moderateComment,
     updateComment,
     likeOrDislikeVideo,
-    videoReactionProcessing,
-    commentInputIsProcessingCollection,
-    processingCommentReactionId,
   }
 }
